requirements.txt
streamlit
pyrebase4
google-cloud-firestore
